requirements.txt
fastapi
uvicorn
requests
python-dotenv
APScheduler
apify-client
python-telegram-bot
ollama